+++ Source/HoudiniEngineEditor/Private/HoudiniEngineEditor.cpp
@@ -1981,7 +1981,7 @@ FHoudiniEngineEditor::PDGBakePackageReplaceModeToPackageReplaceMode(const EPDGBa
 			Mode = FHoudiniPackageParams::GetDefaultReplaceMode();
 			HOUDINI_LOG_WARNING(TEXT("Unsupported value for EPDGBakePackageReplaceModeOption %d, using "
 				"FHoudiniPackageParams::GetDefaultReplaceMode() for resulting EPackageReplaceMode %d"),
-				InReplaceMode, Mode);
+				static_cast<int32>(InReplaceMode), static_cast<int32>(Mode));
 		}
 	}
